+++ src/main/java/br/com/oticanewlook/oticanewlook/model/TipoFuncionario.java
@@ -12,23 +12,23 @@ public class TipoFuncionario {
     
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id_tipo_func;
+    private int id_tipo_func;
 
     private String nome;
 
     public TipoFuncionario() {
     }
 
-    public TipoFuncionario(Long id_tipo_func, String nome) {
+    public TipoFuncionario(int id_tipo_func, String nome) {
         this.id_tipo_func = id_tipo_func;
         this.nome = nome;
     }
 
-    public Long getId_tipo_func() {
+    public int getId_tipo_func() {
         return id_tipo_func;
     }
 
-    public void setId_tipo_func(Long id_tipo_func) {
+    public void setId_tipo_func(int id_tipo_func) {
         this.id_tipo_func = id_tipo_func;
     }
